#43 Refactor
