--- pygeppetto/services/data_source/templates/neo4j.vm
@@ -1,11 +1,0 @@
-{  
-  "statements":[  
-    {  
-      $QUERY,
-      "resultDataContents":[  
-        "row"
-      ],
-      "includeStats":false
-    }
-  ]
-}
